Allow explicit override of Host header.
Allows a default Host header to be set in the `testoverride` section of
the config:

```
---
testoverride:
  dest_addr: some.example.org
  input:
    headers:
      Host: another.example.org
```

It will take precedense over `dest_addr` if it is specified as well. It
will only be set if the input stage does not already have a host header
set.

Falling back to using the DestAddr override if it's not specified.