feat: VirtualHostMode determines `Host` header of internal requests
By default, internal requests (i.e., requests for marking and flushing
the web server log) use `localhost` for the `Host` header value. In
scenarios, where tests are run against a virtual host this will cause
the internal requests to be processed by the default virtual host while
the test requests will be processed by the targeted virtual host. If the
web server logs are segregated by virtual host, internal and test
requests will end up in different log files and go-ftw will not be able
to find the markers it is looking for.

With `virtual_host_mode: true`, the value of the `Host` header from the
test input will be used for internal requests, to ensure that internal
requests will also be processed by the targeted virtual host.

Fixes #361