chore: refactor runner configuration
The `check` package is only used from the `runner` package. It makes
sense to use `TestRunContext` and other types in `check` but that is not
possible due to circular imports. Make `check` part of `runner` instead.

In addition, to reduce issues with cyclic imports and confusion between
different configuration objects, the `RunnerConfig` now consolidates all
configuration data for the test runner.